6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,15 +1,17 @@
 #include <stdio.h>
+
 /**
  * main - print message  code c
  *
  * Return: 0 always runs
  */
+
 int main(void)
 {
 printf("Size of a char: %i bytes(s)\n", sizeof(char));
 printf("Size of an int: %i bytes(s)\n", sizeof(int));
-printf("Size of a long int: %i bytes(s)\n", sizeof(int));
-printf("Size of a long long int: %i bytes(s)\n", sizeof(long));
+printf("Size of a long int: %i bytes(s)\n", sizeof(long));
+printf("Size of a long long int: %i bytes(s)\n", sizeof(long long));
 printf("Size of a float: %i bytes(s)\n", sizeof(float));
 return (0);
 }
